+++ src/test/java/com/island/simcity/SimCityRunnerTest.java
@@ -1,0 +1,10 @@
+package com.island.simcity;
+
+import org.junit.jupiter.api.Test;
+
+public class SimCityRunnerTest {
+    @Test
+    void runSimulation() {
+        SimCityLauncher.main(new String[0]);
+    }
+}